add tool to scrape all queries from grafana
This adds a python utility to scrape the grafana API and print all
PromQL queries, along with the dashboard and panel that the query came
from.

I thought this might be useful when we're looking for usages of a
label, or if we discover a problematic query pattern and want to find
usages of it.