migrate to re-secrets
We have moved our secrets from
alphagov/reliability-engineering-credentials to
alphagov/re-secrets (see alphagov/re-secrets#2).  We should now fetch
our secrets from the new repo rather than the old.

I'd really like to have got rid of the `store_dir` line (and use the
default $PASSWORD_STORE_DIR value) but this only works if there is a
`~/.password-store/.gpg-id` file - and fails in a nonobvious way if
this file is not present.  So, at least for the moment, let's continue
to specify store_dir as a subdirectory of the whole.