Move security group rule to enclave projects
This security group rule caused a dependency cycle between project statefiles:
infra-security-groups depended on the enclave project which in turn depended on
the infra-security-groups project.  This means that we couldn't create a new
environment from scratch.

We fixed this by moving the security group rule into the enclave projects.

Longer term, this rule should be extracted out into a module to DRY up the code.