Add a development enclave for `dm-test-stack` environment
Everything has been hardcoded to work with an equivalent ECS stack
by the same name.

If other devs want development enclaves for their environments then
they should do similar for the moment. In time we might try
and come up with a better solution that having a file commited
for every team member but we have gone with the quickest solution
here.