Point traffic for prom-3 to our third EC2 prom instance
Traffic is routed by nginx based on a new internal
DNS record for EC2 prom-3.

Before this commit each ECS prom was scraping itself and
the other two ECS proms; each EC2 prom was scraping
itself and the other two EC2 proms.

Now ECS prom-1, ECS prom-2 and EC2 prom-3 are each scraping
themselves and each other.

Note, this commit has only added the functionality to the
`dm-test-stack` environment but this will be added to other
paas enclaves in a later commit.

We are only swapping traffic for prom-3 as an incremental step
to make sure everything works before we swap traffic for prom-1
and prom-2.