Use `prom-ec2-x` private DNS for prometheus node exporter targets
We were previously using the EC2 instance private DNS, for
example `ip-10-0-101-79.eu-west-1.compute.internal` as part of the
address for our prometheus node exporter targets. The downside of
this is that you could not tell which prometheus it referred to,
for example is it prom-1, prom-2 or prom-3. By swapping to use
the FQDN's produced by our private zone records we can get
prettier addresses, for example
prom-ec2-1.dm-test-stack.monitoring.private

As we want this consistently across all our stacks that monitor the
PaaS (including dev stacks) there is no longer a need to have
`prometheus_dns_nodes` as a variable for the paas-config` module.

We can also name our variable that get's passed into the template
more accurately as it is not a DNS name but a full address.