Add a check when creating buckets as we are seeing problems with deploys
Problem experienced when deploying to staging or production enclave, you might get an error message like this -

`An error occurred (BucketAlreadyExists) when calling the CreateBucket operation: The requested bucket name is not available. The bucket namespace is shared by all users of the system. Please select a different name and try again.`