Optionally send prometheus syslog to logstash using filebeat
Adds functionality to the prometheus module to send the syslog
of the EC2 instance to logstash using filebeat.

This behaviour is optional as not all stacks will need to send logs
for example dev stacks or the Verify enclave.

We have gone for the most simple configuration for filebeat that
ships syslog but does nothing fancy like adding/editing log lines
or only selecting certain levels. We chose those to keep this work
as small as possible but it may be a good idea to look into
editing this in the future. The downside of this is that log lines
don't explicitely tell you which instance of prometheus they are from.
They do however tell you the private IP address which can be
looked up in the EC2 console to give you the corresponding number
prometheus.