Allow non paas logs to be consumed without parse failure
Our EC2 logs are in a different format to the PaaS logs.
They are loglines from log/syslog rather than the syslog protocol.

We have gone for the quickest and easiest solution which is not
to try and apply any filters to them. They are sufficiently
readable at the moment.