enable session manager access to ECS instances
Just like #190 but for ECS instances.  I suspect that, because we are
using an Amazon Linux AMI and we track the latest changes, we just
need to rebuild our instances to the latest AMI image to get the
latest SSM agent running, so all we need to do is grant SSM
permissions in IAM to the EC2 instance profile.

Once this is in place and working, we should be able to roll back our
use of ssh keys, jumpboxes, and other such things.  That should
hopefully remove a bunch of code.