remove verify-perf-a
We're not using it, and it's just extra stuff for autom8 to think
about during their ubuntu bionic upgrades.  I chatted with @tlwr and
he agrees that we should park this work.

It's in the git history, we can always get it back if we want it :)

My edit to the "enclave" README is a bit of a bodge job.  We can maybe
do better here.