disable ssh in paas-{staging,production} prom EC2
Now that we have Session Manager, we don't need ssh access, so we can
close these security group rules and remove these ssh keys.

We still need ssh support in tests for inspec to gain access to check
resources on the machines, so I've left the `enable_ssh` variable as
configurable and left it to `true` in tests.