hotfix: move prometheus to the private subnets
When we disabled ssh, we also caused these instances to no longer
assign a public IP address.  That caused problems for routing because,
if you're on the public subnet, traffic is routed directly to the
internet gateway and not via a NAT.

The fix is to move the prometheus boxes to the private subnet, where
they should have been anyway.