Change eu-west-2 to eu-west-1
We no longer deploy prometheus to verify-perf-a in eu-west-2
and therefore we no longer need to run kitchen tests in
eu-west-2 either. Everything now should run in eu-west-1 so our
code should reflect that.

I've grepped the code for 'eu-west-2' and also 'london' and these
are the only things I could find.