ensure all DGU alerts keep their org and space labels
To enable #154 to happen, this commit edits the data.gov.uk alerts to
use `without` not `by` clauses, so that we don't unintentionally
remove the `org` and `space` labels everywhere.

The advantage of doing this is that we can then use those labels for
alert routing, and we don't need a separate `product` label at all,
which we can remove at a later date.

The disadvantage is we potentially become a bit noisier about the
labels we're removing - but sometimes explicit is better than
implicit.