fix kitchen tests
3 fixes:

 - typo in a format string in main.tf
 - missed directory change from projects/ to modules/ in prometheus.tf
 - removed test on explicit AMI

The third one deserves some explanation: we used to specify a specific
AMI, but that meant we would not keep up to date with updated
ubuntu images from canonical.

Now that we use the `ami` module to search for the most recent
official canonical AMI, this test doesn't feel useful.  It's
particularly not a test of the `prometheus` module which just uses
whatever AMI you tell it to - if anything, the only thing I think we
could test is whether the common/ami module really does return an AMI
which is an official ubuntu one.  But I think that'd be overkill.