rename "enclave" to "prom-ec2"
We've all agreed that the name "enclave" is not helpful any more,
especially now the actual enclave (in verify) is no more.

I've changed all references to "enclave", except for the statefile
buckets for dm-test-stack, paas-staging and paas-production, where it
seems like a bit of a hassle.

`terraform plan` shows no changes after this commit.