use cf_app_instance arg
This is a retry of #230.  #230 failed because it took the URL-encoded
query parameter and passed it through as the header to cloud foundry.
Because gorouter didn't recognise the instance identifier when it had
`%3A` in place of `:`, it returned a 404 and a failed scrape.

This fix is fugly -- we explicitly search for the `%3A` string -- but
I can't find a way to just get the unescaped value of a query argument
in nginx.  :(