Add auth proxy onto Prometheus EC2 instance
Adds another nginx vhost for the auth proxy. The auth proxy is the same
in functionality to the one that we will be replacing currently running
in ECS.

Removes the default vhost as this would also try and run on port 80.

The vhosts file is fairly different from the original auth-proxy
vhosts file because previously we needed to have different
locations for each of prom 1-3 and alertmanager 1-3 but because
this lives on the box it does not need to be aware of which application
it is fronting and always proxies traffic to a single location -
port 9090 where Prometheus is exposed.

We had to change the password that is being used for basic auth. On
nginx-alpine docker image it came with bcrypt but our ubuntu image
does not. We solved this problem by using a SHA-512 hashing algorithm.
We took this opportunity to pick a new password, generate it's
htpasswd and store it in the re-secrets password (for added security
given that SHA-512 may not be as secure as bcrypt so we won't put
it into a public repo).

Note, this commit does not actually start any user requests going
through the auth-proxy running on the EC2 instances but is a halfway
step of getting there. The later step will be to redirect user
traffics to port 80 on the instance with appropriate changes to
load balancers and DNS.