manage security group for prometheus ALB in this module
not in infra-security-groups.  It's just been confusing having
security groups kept separate from the things they protect.

Adding the security group as an extra group at this stage; we need to
go back and remove monitoring_external_sg_id later.

We avoid having to pass vpc_id in as a separate variable by noting
that we have to be in the same VPC as the subnets we're attaching to.