Swap alertmanager traffic to use new alertmanager ALB
Actually changes traffic over for alertmanager to go through the
alertmanager ALB. Traffic is locked down at the load balancer
using security groups to office IPs.

After this is deployed we can then remove the nginx auth-proxy in
ECS and related load balancers/security groups.

Alertmanager itself remains in ECS.

Note, this will cause downtime between the deploy of the
ALBs and the alertmanager ECS service whilst the route53
records are swapped but the alertmanagers are not yet
registered to the new target groups.