Allow alertmanager ALB to send traffic to alertmanager
We need this rule to allow egress from the alertmanager ALB to
the ECS instances running alertmanager and ingress to accept
it.

Note, we can't do this by adding the `external-monitoring` sg
to our alertmanager ALB (which would allow it to egress traffic
to the internal-monitoring sg) because then we would also open
up the alertmanager ALB to all public internet traffic.