Simply alerts for prom and alertmanager thresholds
These alerts were not acting the way we expected. They would fire
if a single prometheus or alertmanager went down but we wanted
it to fire only if two or more had gone down. This was causing issues
with falsely firing alerts during deployments.

I spent 5 minutes fiddling with the alert expression to maintain
the current behaviour of identifying which individual instance was
down but in the end decided that it is not worth the time or
upkeep. It is a 30 second job to identify which instance is having
the problem by visiting the web URLs.

Maybe when we have alert unit testing adding we can add this
behaviour back in but for the moment I'm going to keep our alerts
as simple as possible to be confident in their behaviour.