Add relabel_configs to allow /_metrics scrape paths for DM
Digital marketplace are instrumenting their apps, and they have URL
slugs that are dependent on content submitted by users.  This means
that `/.../metrics` isn't guaranteed to be available on their apps (and in
fact is not available on one of them).

They are able to serve metrics from a `/.../_metrics` endpoint
instead.  But I can't think of any general way for them to tell
us (via the service broker) about a custom metrics endpoint that we
should be scraping.

In lieu of a clean, self-service solution like that, I offer this hack
instead.  This commit adds relabel_configs to the paas-targets scrape
config.  They detect any apps running in the `digitalmarketplace` org,
with app name `api`, `*-api` or `*-frontend`, and change the
`__metrics_path__` label to end in `/_metrics` instead of `/metrics`.

Targets which do not match the regex will be unaffected and continue
to work as before.