configure alertmanager using the entrypoint hack
This passes alertmanager's configuration in to the task as a
base64-encoded environment variable.  We then override the container's
entrypoint to insert some shell scripting that writes the environment
variable to a file, and finally starts alertmanager.

We still use terraform template_files to fill in the alertmanager
config details, but we could in future do the templating in the
entrypoint using envsubst or similar, which would mean we could keep
our secrets in parameter store rather than in the task definition.

The old way was pretty awful, which means i get to delete a lot of
code.  I'm not deleting the S3 config bucket in this commit, just in
case, but it should be fine to delete once this has been deployed.