Don't install so many packages
If you do a plain `apt-get install` you get loads of stuff shoveled in
as "Recommended" dependencies.  The `--no-install-recommends` option
tells apt to stop that.

As a result, we need to explicitly pull in gpg and gpg-agent.

This reduces the number of packages we install from 72 to 14, and
reduces overall build times from 2 minutes to 1 minute 30.