[alerts] add group_by
This adds a group_by config item.  The main thing is that I'd like us
to group_by `alertname` so that alerts with different names get
different notifications in slack.

This might be something that we only want to set on slack receivers
rather than for all receivers, but for now I'd rather keep the
configuration simple and we can tweak it later if it's not ideal.