Change how we match GSP alerts
The problem with matching on `deployment: gsp` is that we apply
`deployment: gsp` as an external label.  But external labels are only
applied when the underlying alert does not already have the
`deployment` label.  And Kubernetes has things called Deployments, and
we generate alerts for Deployments that have a `deployment` label that
identifies the Deployment.

So, these alerts have been finding their way into the verify-2ndline
receiver, and confusing people.

As a (hacky) workaround, I'm matching on the `clustername` label
instead.  For verify, it'll be `london.verify.govsvc.uk`; I'm assuming
that all other clusternames will be of a particular form.  That's a
bad assumption and we should find a better solution.  But this commit
will fix the immediate problem.