Correct parenthesis grouping
The parentheses on this were off which meant in such a way that they
parsed correctly(!) but didn't do what was expected.

This change fixes the grouping and tries to be more explicit about where
some of the magic numbers come from.

The goal of this alert is to predict disks that will be full in the next
3 days but only on nodes that have been up for the last 12 hours. We
ignore new instances from the prediction because the prediction often
causes false positives for new instances.