route "layer" label to #re-autom8-alerts only
the re-autom8-alert channel is so noisey it is unusable. Instead of
routing ALL the cluster alerts to the channel we will only route things
to the channel that have an explicit "layer" label.

The "label" label is already used elsewhere to denote a grouping of
components or a system of components.

For example:

`layer: cicd` would signal that the alert is related to the continuous
deployment tooling/systems

`layer: orchestration` could signal that the alert is related to the
scheduling and execution of containers

`layer: routing` might signal that the alert is related to ingress,
networking or load-balancing

right now we haven't started using this label, so we route any alert
that has not matched a previous receiver and has this label present to
`#re-autom8-alerts` slack channel.