Spike: demonstrate prometheus-buildpack
So I've been playing around with running prometheus on the PaaS using
the influxdb private beta.  I've got something working over at
alphagov/prometheus-buildpack and this commit shows how to use it.

The vague idea is that eventually we might be able to migrate Observe
Prometheus to run in PaaS itself, with influxdb as the backing store.
This commit gets us:

 - Prometheus running on PaaS
 - Using influxdb as a remote_write/remote_read endpoint for long-term
   storage

Things not yet done that we might want in order to migrate Observe to
PaaS:

 - A [sidecar][] to pull targets from an S3 bucket
 - Migrate to the [S3 broker][] (here and in cf_app_discovery)
 - An answer for how we do access control from Prometheus to the
   target apps (currently it's either via basic auth or via IP
   safelisting)
 - Consider moving away from the service broker to something else
   entirely (such as using [labels and selectors][] to discover apps
   to scrape)

The last two points are possibly interdependent.  In particular, when
running Prometheus on PaaS, we could choose to scrape targets over
private networking rather than public.  This neatly sidesteps the
access control issue, but raises another issue of how we enable
cross-org network security policies.  The problem is that a single
cloud foundry user needs SpaceDeveloper permission in both the space
where Prometheus runs and the space where the scrape target is
running.  If we were to go down this route, I'd imagine we'd want to
ditch the service broker in favour of a fresh labels-and-selectors
approach.

Alternatively, we can continue to scrape over the public internet.
The problem here is that we predominantly use IP safelisting, but this
is probably much less appropriate for a Prometheus running in PaaS,
which will share the same egress IPs as everything else running on
PaaS.

[labels and selectors]: https://v3-apidocs.cloudfoundry.org/version/3.78.0/index.html#labels-and-selectors
[S3 broker]: https://docs.cloud.service.gov.uk/deploying_services/s3/
[sidecar]: https://docs.cloudfoundry.org/devguide/sidecars.html