[app-ecs-services] better tagging
"Environment" is more meaningful than "Stack name" and it's also one
of our billing tags.

I also tagged more resources than we used to tag:

 - the ECS cluster itself
 - the IAM User for SMTP
 - the IAM Role for the alertmanager task execution
 - task definitions

I chose not to tag ECS services, because they require opting in to the
new ECS ARN format and I'm not sure if/how to do that right now.