Use non-conflicting CN for old ALBs
We're replacing the ALB+certificate with a freshly issued certificate on
an NLB.

We cannot issue separate certificate for the same CN (`alerts.domain`)
without also creating conflicting DNS entries required for domain
validation. This change works around this and can be removed once we get
round to destroying the old ALB infrastructure.