Rename `app-ecs-services` to `alertmanager`
This renames the `app-ecs-services` module to `alertmanager` because
that's now what it does and it's a much clearer name.

This commit needs to be deployed carefully, because it renames the
app-ecs-services modules.  We need to pause the concourse pipeline,
then run

    terraform state mv module.app-ecs-services module.alertmanager

for each of terraform/projects/alertmanager-{staging,prod}.  Then we
can merge the PR, apply the new pipeline.yml, and unpause the pipeline
to have it do the deploy.