[alertmanager] create separate ECS service per AZ
Currently we have a single ECS service with desired_count = 3 and we
hope that the tasks will be launched one per AZ/subnet.  However this
isn't the behaviour we see.

This changes things so that we have three separate services, with a
desired count of 1 each, and each in a separate subnet (and hence
separate AZ).  The service names have the AZ name in them to make it
clear which one is which.