Scrape PaaS's alpha Elasticsearch metrics endpoint
PaaS have deployed a new metrics endpoint which collects all metrics
for Digital Marketplace's Elasticsearch instances. This commit adds
the new scrape target to the Prometheus config and uses a `pass`
secret as the password for the scrape endpoint.

The scrape config for this endpoint is very large because there are
700+ metrics exposed and we only want to import metrics which we
expect to be useful. Future versions of the Elasticsearch exporter
will try to improve on that.