Allow extra env-specific Prometheus configs
This commit adds an `extra_scrape_configs` input variable to the
`prom-ec2/paas-config` Terraform module. Terraform data can be passed
to it representing a list of extra Prometheus `scrape_configs` [1].
Typically this will come by running `yamldecode()` on YAML from a
file.

This commit then utilises the feature to move the Digital Marketplace
Elasticsearch metrics out of the generic module and into the
production deployment of Prometheus.

[1] https://prometheus.io/docs/prometheus/latest/configuration/configuration/#scrape_config