Make RE_Observe_No_Paas_Targets a bit less trigger-happy
This alert has been quite noisy recently, and it's only there to catch
a logical misconfiguration (rather than a reliability problem caused
by some sort of transient failure).

I think we can bump up the `for` clause to make it less trigger-happy.
It will still fire in a reasonably timely manner if we somehow mess up
the service broker configuration, but won't fire randomly just because
we deployed some code and rebuilt the prometheus servers.