Point per-az dns names at alertmanager alb
This directs the per-az DNS names such as
`alerts-eu-west-1a.monitoring.gds-reliability.engineering` at the ALB,
not the NLB.

They won't be in use yet (by prometheus, anyway) because prometheis
are currently set up to discover A records on `alerts.<domain>`.  But
this provides us a migration path from old NLB discovery to new ALB
discovery.