observe alerts: add RE_Observe_No_Successful_Updates alert
this is an alert we want to fire when we haven't seen any successful
targets updates in a region for a certain period. updates are triggered
through the /update-targets endpoint, so luckily we should just be able to
look at the 200 responses to these requests to see successful updates.

wrap in an outer sum() aggregation to leave open the possibility of
multiple app instances