Don't pagerduty for metric exporters or conduit disk space
We wouldn't be losing any critical user functionality if one of these
apps were to fill up with disk space. We will get the 75% ticket alert
for disk space for these apps but we don't need to wake someone up
because of them as it does not require immediate action like the other
apps.