Add p2 pagerduty rule for notify and use for 95% disk space
p2 has been chosen for disk space filling up because the likelyhood is
that if disk space fills up to 100% then it will not be a p1 outage as
the user impact will not be a full outage. Therefore it is a p2 as it
needs urgent attention and may be affecting some componenets. Note, if
indeed the disk filling up to 100% does cause a p1 type outage, our
other alerts/smoke tests will cover this to make sure people are
alerted/woken up.