ci: consolidate prometheus-aws-configuration-beta git resources
going from 4 to 2. most jobs are fairly fast when applying an empty change
and having many split git resources creates a bunch of weird race
conditions. one task that is expensive for every run is building the
task image, so move that to its own job (from the configure-pipeline job)
with its own dedicated git resource. this way simultaneous changes
to the task image and the rest of the repo should be the only case that
can introduce weird race conditions.